% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{EM}
\alias{EM}
\title{EM algorithm}
\usage{
EM(W,nsteps,a,b,eps)
}
\arguments{
\item{W}{(KxK) matrix where the (i,j) element is number of matches player i has beaten player j.}

\item{nsteps}{number of iterations.}

\item{a}{shape parameter for Gamma prior.}

\item{b}{rate parameter for Gamma prior.}

\item{eps}{threshold for change in lambda.}
}
\value{
The final estimates of lambdas.
}
\description{
Performs the update from function EM_step_step n times.
A threshold is defined such that the function will break if the change in lambda is smaller than the threshold value.
}
