% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{EM_step}
\alias{EM_step}
\title{One step of EM algorithm}
\usage{
EM_step(lambda,W,a,b,K,n)
}
\arguments{
\item{lambda}{current values for the estimates of lambda.}

\item{W}{(KxK) matrix where the (i,j) element is number of matches player i has beaten player j.}

\item{a}{shape parameter for Gamma prior.}

\item{b}{rate parameter for Gamma prior.}

\item{K}{number of players.}

\item{n}{W + t(W) - (i,j) element is total number of times player i has played player j.}
}
\value{
Updated estimate of lambdas.
}
\description{
One step of EM algorithm
}
