% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bradley_terry.R
\name{extract_score}
\alias{extract_score}
\title{Find the predicted Bradley-Terry score of a player}
\usage{
extract_score(player,lambda_hat,players)
}
\arguments{
\item{player}{id of the player for whom we would like the estimated Bradley-Terry score.}

\item{lambda_hat}{vector contain a list of all the predicted scores from the EM algorithm.}

\item{players}{vector containing a list of players for which there is a predicted Bradley-Terry score.}
}
\value{
The predicted score of the player. If the player is not found within the list of players then -1 is returned.
}
\description{
Finds the predicted score of a player following the use of the EM algorithm. If no such score exists for a player -1 is returned as the score for that player.
}
