% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_matrix.R
\name{feature_surface}
\alias{feature_surface}
\title{Matrix of surface features}
\usage{
feature_surface(my_data, no_matches, player_data, p1_p2, surface, carpet = 'FALSE')
}
\arguments{
\item{my_data}{input data.}

\item{no_matches}{number of matches played.}

\item{player_data}{player data frame which is the output of function player_data_frame.}

\item{p1_p2}{a matrix where each column signifies a match and the two rows are the player ids for player 1 and player 2 for that match.}

\item{surface}{a list of surfaces played on.}

\item{carpet}{as matches are no longer played on carpet, carpet performance is by default not included. However, carpet can be included by including, carpet == 'TRUE'.}
}
\value{
A matrix of players surface features depending on which surface the match is played on.
}
\description{
Let n be the number of surfaces and m the number of matches . Creates a (2n x m) matrix where the first n rows describe surface features for player 1 and the next n rows describe surface features for player 2.
Each column signifies a match and the only non-zero elements of that column are for the 2 rows (one for each player) which describe the surface on which the match was played.
The two values for the column will be the performance of player 1 and player 2 respectively on that surface.
}
\examples{
feature_surface(data, no_matches = 1000, player_data, p1_p2, c("Hard","Clay", "Grass"))
feature_surface(data, no_matches = 31205, player_data, p1_p2, c("Hard","Clay", "Grass", "Carpet"), carpet = 'TRUE')
}
