% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_matrix.R
\name{head_to_head}
\alias{head_to_head}
\title{Head to head performance}
\usage{
head_to_head(my_data, player1, player2)
}
\arguments{
\item{my_data}{input data.}

\item{player1}{player 1 id.}

\item{player2}{player 2 id.}
}
\value{
Vector with two values, first, the percentage of games player 1 beat player 2 and second, the percentage of games player 2 beat player 1.
}
\description{
Calculates the number of matches won/loss against each other. However, if players have never played each other both outputs are set to 0.5.
}
