% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processed_data.R
\name{load_data}
\alias{load_data}
\title{Load data}
\usage{
load_data(file_list, variables = FALSE)
}
\arguments{
\item{file_list}{a list of files to be loaded.}

\item{variables}{the variables from the files that we wish to load.}
}
\value{
A data frame with columns for each variable and each row is a match.
}
\description{
Loads multiple files of data and extracts features that are desired. It also formats the date and puts data in date order.
}
\examples{
load_data("wta_matches_2000.csv")
load_data(file_list, variables = c('tourney_date','surface', 'winner_id', 'loser_id', 'winner_age', 'loser_age'))
}
