% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_reg.R
\name{log_reg}
\alias{log_reg}
\title{Logistic regression}
\usage{
log_reg(x,y)
}
\arguments{
\item{x}{feature matrix for training data.}

\item{y}{output for training data.}
}
\value{
The estimated weight parameters.
}
\description{
Define a function that returns the estimated weight parameters using logistic regression.
}
