% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_reg.R
\name{log_reg_naive}
\alias{log_reg_naive}
\title{Likelihood function}
\usage{
log_reg_naive(w, x, y)
}
\arguments{
\item{w}{matrix where the (i,j) element is number of matches player i has beaten player j.}

\item{x}{feature matrix for training data.}

\item{y}{output for training data.}
}
\value{
The negative log-likelihood.
}
\description{
Define the negative log-likelihood function to be maximised.
}
