% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_matrix.R
\name{match_based_form_total}
\alias{match_based_form_total}
\title{Historical performance of player}
\usage{
match_based_form_total(my_data, player)
}
\arguments{
\item{my_data}{input data.}

\item{player}{player id.}
}
\value{
The percentage of total matches won.
}
\description{
Calculates the percentages of matches won from all matches in the data set.
}
