% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_matrix.R
\name{player_data_frame}
\alias{player_data_frame}
\title{Construct player data frame}
\usage{
player_data_frame(my_data)
}
\arguments{
\item{my_data}{input data.}
}
\value{
A data frame with columns: player_id, age, one for each surface and form.
}
\description{
Creates a data frame of all players using functions surface_based_form_total and match_based_form_total.
}
