% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_analysis.R
\name{plot_comparisons}
\alias{plot_comparisons}
\title{Plots graph of matrix}
\usage{
plot_comparisons(w)
}
\arguments{
\item{w}{matrix to be plotted.}
}
\value{
The graph of matrix.
}
\description{
Plots the graph formed by a matrix.
}
