% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bradley_terry.R
\name{predict_winner}
\alias{predict_winner}
\title{Predict the winner of a series of matches}
\usage{
predict_winner(test_dat,lambda_hat,players)
}
\arguments{
\item{test_dat}{a matrix containing the matches played.}

\item{lambda_hat}{vector contain a list of all the predicted scores from the EM algorithm.}

\item{players}{vector containing a list of players for which there is a predicted Bradley-Terry score.}
}
\value{
The probability that player 1 beats player 2 using the Bradley-Terry formula: lambda_1/(lambda_1+lambda_2). If either of the player does not have a predicted score then 0.5 is returned to express uncertainty.
}
\description{
Finds the probability of player1 winning in each match, using the predicted Bradley-Terry scores.
}
