% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_analysis.R
\name{reduce_to_connected}
\alias{reduce_to_connected}
\title{Forms strongly connected graph}
\usage{
reduce_to_connected(w)
}
\arguments{
\item{w}{matrix to be reduced.}
}
\value{
The reduced matrix.
}
\description{
Reduces a matrix to one that forms a strongly connected graph.
}
