% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_reg.R
\name{sigma_f}
\alias{sigma_f}
\title{Sigmoid function}
\usage{
sigma_f(t)
}
\arguments{
\item{t}{input to the sigmoid function.}
}
\value{
Output of sigmoid function.
}
\description{
Define the sigmoid function.
}
