% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_analysis.R
\name{strongly_connected}
\alias{strongly_connected}
\title{Check for strongly connected graph}
\usage{
strongly_connected(w)
}
\arguments{
\item{w}{matrix to check.}
}
\value{
A statement that is either: 'The graph is strongly connected' or if the graph is not strongly connected then:
'The graph contains n strongly connected components and the largest one contains m nodes' where n and m will be determined.
}
\description{
Determines whether a matrix forms a strongly connected graph.
}
