% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_matrix.R
\name{surface_based_form_total}
\alias{surface_based_form_total}
\title{Performance of player on different surfaces}
\usage{
surface_based_form_total(my_data, player, surface)
}
\arguments{
\item{my_data}{input data.}

\item{player}{player id.}

\item{surface}{a list of surfaces played on.}
}
\value{
A list with percentages of matches won on that surface.
}
\description{
Calculates percentage of matches won by a player on different surfaces.
}
