% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bradley_terry.R
\name{w_fill}
\alias{w_fill}
\title{Construct the matrix of players wins/losses}
\usage{
w_fill(data_mat, players)
}
\arguments{
\item{data_mat}{table containing matches with variables winner_id and loser_id.}

\item{players}{vector containing a list of players from which to construct the matrix.}
}
\value{
A square matrix with dimension equal to the number of players
}
\description{
Constructs a matrix with (i,j) which denotes the number of matches where player i beats player j.
}
